+++ src/OmniSharp.MSBuild/Logging/MSBuildDiagnostic.cs
@@ -1,5 +1,9 @@
-﻿namespace OmniSharp.MSBuild.Logging
+﻿using System;
+using OmniSharp.Utilities;
+
+namespace OmniSharp.MSBuild.Logging
 {
+
     public class MSBuildDiagnostic
     {
         public MSBuildDiagnosticSeverity Severity { get; }
@@ -31,13 +35,43 @@
         }
 
         public static MSBuildDiagnostic CreateFrom(Microsoft.Build.Framework.BuildErrorEventArgs args)
-            => new MSBuildDiagnostic(MSBuildDiagnosticSeverity.Error,
-                args.Message, args.File, args.ProjectFile, args.Subcategory, args.Code,
+        {
+            string diagnosticText = null;
+
+            // https://github.com/dotnet/msbuild/blob/v16.8.3/src/Tasks/Resources/Strings.resx#L2155-L2158
+            // for MSB3644, we should print a different message on Unix because the default one is Windows-specific
+            if (args.Code.Equals("MSB3644", StringComparison.OrdinalIgnoreCase))
+            {
+                // https://github.com/dotnet/msbuild/issues/5820
+                // older versions of MSBuild incorrectly treat 'net5.0'/'net6.0' moniker as ".NETFramework,Version=v5.0/6.0"
+                // this creates a confusing error message which we convert into a more helpful one
+                if (args.Message.Contains(".NETFramework,Version=v5.0"))
+                {
+                    diagnosticText = Platform.Current.OperatingSystem == Utilities.OperatingSystem.Windows
+                    ? ErrorMessages.ReferenceAssembliesNotFoundNet50Windows : ErrorMessages.ReferenceAssembliesNotFoundNet50Unix;
+                }
+                else if (args.Message.Contains(".NETFramework,Version=v6.0"))
+                {
+                    diagnosticText = Platform.Current.OperatingSystem == Utilities.OperatingSystem.Windows
+                    ? ErrorMessages.ReferenceAssembliesNotFoundNet60Windows : ErrorMessages.ReferenceAssembliesNotFoundNet60Unix;
+                }
+                else
+                {
+                    diagnosticText = Platform.Current.OperatingSystem != Utilities.OperatingSystem.Windows
+                        ? ErrorMessages.ReferenceAssembliesNotFoundUnix : args.Message;
+                }
+            }
+
+            return new MSBuildDiagnostic(MSBuildDiagnosticSeverity.Error,
+                diagnosticText ?? args.Message, args.File, args.ProjectFile, args.Subcategory, args.Code,
                 args.LineNumber, args.ColumnNumber, args.EndLineNumber, args.EndColumnNumber);
+        }
 
         public static MSBuildDiagnostic CreateFrom(Microsoft.Build.Framework.BuildWarningEventArgs args)
-            => new MSBuildDiagnostic(MSBuildDiagnosticSeverity.Error,
-                args.Message, args.File, args.ProjectFile, args.Subcategory, args.Code,
-                args.LineNumber, args.ColumnNumber, args.EndLineNumber, args.EndColumnNumber);
+        {
+            return new(MSBuildDiagnosticSeverity.Error,
+                           args.Message, args.File, args.ProjectFile, args.Subcategory, args.Code,
+                           args.LineNumber, args.ColumnNumber, args.EndLineNumber, args.EndColumnNumber);
+        }
     }
 }